[REFACTOR] Notice, FAQ 관련 패키지 수정
