--- src/main/java/ject/componote/domain/announcement/api/FAQController.java
@@ -1,4 +1,0 @@
-package ject.componote.domain.announcement.api;
-
-public class FAQController {
-}
