--- src/main/java/ject/componote/domain/announcement/api/NoticeController.java
@@ -1,4 +1,0 @@
-package ject.componote.domain.announcement.api;
-
-public class NoticeController {
-}
