--- src/main/java/ject/componote/domain/announcement/error/NoticeException.java
@@ -1,4 +1,0 @@
-package ject.componote.domain.announcement.error;
-
-public class NoticeException {
-}
