--- src/main/java/ject/componote/domain/announcement/model/Title.java
@@ -1,20 +1,0 @@
-package ject.componote.domain.announcement.model;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-@EqualsAndHashCode
-@Getter
-@ToString
-public class Title {
-    private final String value;
-
-    private Title(final String value) {
-        this.value = value;
-    }
-
-    public static Title from(final String value) {
-        return new Title(value);
-    }
-}
