+++ src/main/java/ject/componote/domain/faq/api/FAQController.java
@@ -1,0 +1,4 @@
+package ject.componote.domain.faq.api;
+
+public class FAQController {
+}
