+++ src/main/java/ject/componote/domain/faq/dao/FAQRepository.java
@@ -1,6 +1,6 @@
-package ject.componote.domain.announcement.dao;
+package ject.componote.domain.faq.dao;
 
-import ject.componote.domain.announcement.domain.FAQ;
+import ject.componote.domain.faq.domain.FAQ;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FAQRepository extends JpaRepository<FAQ, Long> {
