+++ src/main/java/ject/componote/domain/faq/error/FAQException.java
@@ -1,0 +1,4 @@
+package ject.componote.domain.faq.error;
+
+public class FAQException {
+}
