+++ src/main/java/ject/componote/domain/notice/dao/NoticeRepository.java
@@ -1,6 +1,6 @@
-package ject.componote.domain.announcement.dao;
+package ject.componote.domain.notice.dao;
 
-import ject.componote.domain.announcement.domain.Notice;
+import ject.componote.domain.notice.domain.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
