+++ src/main/java/ject/componote/domain/notice/error/NoticeException.java
@@ -1,0 +1,4 @@
+package ject.componote.domain.notice.error;
+
+public class NoticeException {
+}
